Moved floating action button to ProfileFragment.
